Separate main() to main.cpp and add script to compile ecosystem
