+++ src/cpp_version/compile_ecosystem.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+g++ main.cpp ecosystem.cpp --std=c++11 -o ecosystem
